Changes to signandverify tests
