+++ src/UnitTests/SignAndVerify/X509Certificate2_VerifySignature.cs
@@ -56,5 +56,33 @@ namespace UnitTests.SignAndVerify
             Assert.IsTrue(good);
         }
         #endregion
+
+        #region Happy scenarios
+        [TestMethod]
+        public void UsingMultipleHashAlgorithms_ShouldWork()
+        {
+            const string TestData = "Hello world";
+
+            var payload = Encoding.UTF8.GetBytes(TestData);
+
+            string[] HashAlgorithmNames = { "SHA256", "SHA384", "MD5", "SHA1", "SHA512" };
+
+            foreach (var name in HashAlgorithmNames)
+            {
+                using var hashAlgorithm = HashAlgorithm.Create(name);
+                // Digest
+                var hash = hashAlgorithm.ComputeHash(payload);
+                Console.WriteLine($"Hash: {name} {hash.Length * 8} bits / {hash.Length} BYTES");
+
+                // Sign
+                var signature = MyConfig.SigningCertificate.CreateSignature(hash);
+                Console.WriteLine($"Signature: {signature.Length * 8} bits / {signature.Length} BYTES");
+
+                // Verify
+                var good = MyConfig.VerifyCertificate.VerifySignature(hash, signature);
+                Assert.IsTrue(good);
+            }
+        }
+        #endregion
     }
 }
