+++ src/X509.EnduranceTest.Shared/SignAndVerify/X509Certificate2_SignAndVerifyEnduranceTests.cs
@@ -8,9 +8,12 @@ namespace X509.EnduranceTest.Shared
 {
     internal class X509Certificate2_SignAndVerifyEnduranceTests
     {
-        internal static void Run( double dataSizeInKB, int maxIterations)
+        internal static void RunSign(
+            string hashName ,
+            //Below dataSizeInKB doesn't make sense as the hash is signed and hash is fixed size.
+            double dataSizeInKB, 
+            int maxIterations)
         {
-            var hashName = "SHA256";
             X509Certificate2 cert = MyConfig.SigningCertificate;
             var payload = TestDataGenerator.GenerateJunk((int)(dataSizeInKB * 1024));
             Console.WriteLine($"Generated {payload.Length / 1024} KB random binary data.");
@@ -21,10 +24,10 @@ namespace X509.EnduranceTest.Shared
             // Act
             EnduranceTestRunner.Run(maxIterations, () => cert.CreateSignature(hash));
         }
-        internal static void RunVerify(double dataSizeInKB, int maxIterations)
+        internal static void RunVerify(string hashName, double dataSizeInKB, int maxIterations)
         {
             var payload = TestDataGenerator.GenerateJunk((int)(dataSizeInKB * 1024));
-            using var hashAlgorithm = HashAlgorithm.Create("SHA256");
+            using var hashAlgorithm = HashAlgorithm.Create(hashName);
             var hash = hashAlgorithm.ComputeHash(payload);
             var signature = MyConfig.SigningCertificate.CreateSignature(hash);
             X509Certificate2 verifyCertificate = MyConfig.VerifyCertificate;
